When a user account is deleted, the app makes an API call to delete their avatar from Cloudinary

+++ controllers/adminController.js
@@ -213,10 +213,31 @@ exports.deleteUser = async ( req, res ) => {
         res.redirect( 'back' );
         return;
     } else {
-        await User.findOneAndDelete({ _id: req.params.user_id });
-        req.flash( 'success', 'You successfully deleted the user account.' );
-        res.redirect( 'back' );
-        return;
-    }
 
+        const userToDelete = await User.findOne({ _id: req.params.user_id });
+
+        if ( !userToDelete ) {
+
+            req.flash( 'error', 'There was an error finding the user you are attempting to delete. Please try again.' );
+            res.redirect( 'back' );
+            return;
+
+        } else {
+
+            await User.findOneAndDelete({ _id: req.params.user_id });
+
+            await cloudinary.v2.api.delete_resources( userToDelete.avatar_id,
+                function( error, result ) {
+                    if ( error ) {
+                        req.flash( 'error', 'This users avatar could not be deleted. You will have to go and delete it manually.' );
+                    }
+                }
+            );
+    
+            req.flash( 'success', 'You successfully deleted the user account.' );
+            res.redirect( 'back' );
+            return;
+
+        }
+    }
 }
